homend_server/requirements.txt
Django==1.8.4
django-braces==1.8.1
django-cors-headers==1.1.0
django-oauth-toolkit==0.9.0
djangorestframework==3.2.3
docutils==0.12
oauthlib==1.0.1
Pillow==3.0.0
six==1.9.0
wheel==0.24.0
